requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Cors==3.0.10
Flask-WTF==0.15.1
tensorflow>=2.10.0
Pillow>=9.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
h5py==3.3.0
Werkzeug==2.0.1 